Update to SPFx 1.8
