--- samples/react-application-injectcss/lib/extensions/injectCss/InjectCssApplicationCustomizer.d.ts
@@ -1,7 +1,0 @@
-import { BaseApplicationCustomizer } from '@microsoft/sp-application-base';
-export interface IInjectCssApplicationCustomizerProperties {
-    cssurl: string;
-}
-export default class InjectCssApplicationCustomizer extends BaseApplicationCustomizer<IInjectCssApplicationCustomizerProperties> {
-    onInit(): Promise<void>;
-}
