+++ samples/react-application-injectcss/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "react-application-injectcss",
+  "version": "0.0.1",
+  "private": true,
+  "engines": {
+    "node": ">=0.10.0"
+  },
+  "scripts": {
+    "build": "gulp bundle",
+    "clean": "gulp clean",
+    "test": "gulp test"
+  },
+  "dependencies": {
+    "@microsoft/sp-core-library": "1.8.0",
+    "@microsoft/decorators": "1.8.0",
+    "@types/webpack-env": "1.13.1",
+    "@types/es6-promise": "0.0.33",
+    "@microsoft/sp-dialog": "1.8.0",
+    "@microsoft/sp-application-base": "1.8.0"
+  },
+  "devDependencies": {
+    "@microsoft/sp-build-web": "1.8.0",
+    "@microsoft/sp-tslint-rules": "1.8.0",
+    "@microsoft/sp-module-interfaces": "1.8.0",
+    "@microsoft/sp-webpart-workbench": "1.8.0",
+    "@microsoft/rush-stack-compiler-2.7": "0.4.0",
+    "gulp": "~3.9.1",
+    "@types/chai": "3.4.34",
+    "@types/mocha": "2.2.38",
+    "ajv": "~5.2.2"
+  }
+}
