+++ samples/react-application-injectcss/sample2.css
@@ -1,0 +1,4 @@
+/* Changes the feedback button color */
+.sitePage-uservoice-button {
+  background-color: orangered;
+ }
