+++ samples/react-application-injectcss/sharepoint/assets/ClientSideInstance.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Elements xmlns="http://schemas.microsoft.com/sharepoint/">
+    <ClientSideComponentInstance
+        Title="InjectCss"
+        Location="ClientSideExtension.ApplicationCustomizer"
+        ComponentId="5a1fcffd-dfeb-4844-b478-1feb4325a5a7"
+        Properties="{&quot;cssurl&quot;:&quot;/Style%20Library/custom.css&quot;}">
+    </ClientSideComponentInstance>
+</Elements>
