+++ samples/react-application-injectcss/sharepoint/assets/elements.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Elements xmlns="http://schemas.microsoft.com/sharepoint/">
+    <CustomAction
+        Title="InjectCss"
+        Location="ClientSideExtension.ApplicationCustomizer"
+        ClientSideComponentId="5a1fcffd-dfeb-4844-b478-1feb4325a5a7"
+        ClientSideComponentProperties="{&quot;cssurl&quot;:&quot;/Style%20Library/custom.css&quot;}">
+    </CustomAction>
+</Elements>
