+++ samples/react-application-injectcss/src/extensions/injectCss/InjectCssApplicationCustomizer.manifest.json
