+++ samples/react-application-injectcss/src/extensions/injectCss/InjectCssApplicationCustomizer.ts
@@ -1,0 +1,36 @@
+import { override } from '@microsoft/decorators';
+import { Log } from '@microsoft/sp-core-library';
+import {
+  BaseApplicationCustomizer
+} from '@microsoft/sp-application-base';
+
+import * as strings from 'InjectCssApplicationCustomizerStrings';
+
+const LOG_SOURCE: string = 'InjectCssApplicationCustomizer';
+
+
+export interface IInjectCssApplicationCustomizerProperties {
+  cssurl: string;
+}
+
+export default class InjectCssApplicationCustomizer
+  extends BaseApplicationCustomizer<IInjectCssApplicationCustomizerProperties> {
+
+  @override
+  public onInit(): Promise<void> {
+    Log.info(LOG_SOURCE, `Initialized ${strings.Title}`);
+
+    const cssUrl: string = this.properties.cssurl;
+    if (cssUrl) {
+        // inject the style sheet
+        const head: any = document.getElementsByTagName("head")[0] || document.documentElement;
+        let customStyle: HTMLLinkElement = document.createElement("link");
+        customStyle.href = cssUrl;
+        customStyle.rel = "stylesheet";
+        customStyle.type = "text/css";
+        head.insertAdjacentElement("beforeEnd", customStyle);
+    }
+
+    return Promise.resolve();
+  }
+}
