+++ samples/react-application-injectcss/src/extensions/injectCss/loc/en-us.js
