+++ samples/react-application-injectcss/src/extensions/injectCss/loc/myStrings.d.ts
@@ -1,0 +1,8 @@
+declare interface IInjectCssApplicationCustomizerStrings {
+  Title: string;
+}
+
+declare module 'InjectCssApplicationCustomizerStrings' {
+  const strings: IInjectCssApplicationCustomizerStrings;
+  export = strings;
+}
