+++ samples/react-application-injectcss/src/index.ts
@@ -1,2 +1,1 @@
 // A file is required to be in the root of the /src directory by the TypeScript compiler
-//# sourceMappingURL=index.js.map
